Ilmor: use trim_adjusted for soft-stop

+++ libraries/AP_Ilmor/AP_Ilmor.h
@@ -88,7 +88,7 @@ private:
     AP_Int8 _can_port;
     AP_Int16 _trim_stop;
 
-    uint16_t _current_trim_position;
+    uint8_t _current_trim_position;
     uint8_t _trim_command_from_buttons;
     bool _trim_locked_out = true;
     uint32_t _last_wait_ms;
